+++ sources/Panel/src/Menu/Pages/PageSignals.h
@@ -31,7 +31,7 @@ static ChoiceParameter cpParameters =
 
 void PanelSignals_Init()
 {
-    waveForm = WAVE_FORM;
+    waveForm = (uint8)WAVE_FORM;
     InputWindowStruct::FillAllowParameters(CHANNEL, WAVE_FORM, &cpParameters.allowParameters);
     cpParameters.numParameter = (uint8 *)&(set.sig_parameter[(WaveForm)waveForm]);
 }
@@ -90,7 +90,7 @@ static void OnPress_Form(void)
 
 static void OnPress_Channel(void)
 {
-    waveForm = WAVE_FORM;
+    waveForm = (uint8)WAVE_FORM;
     InputWindowStruct::FillAllowParameters(CHANNEL, WAVE_FORM, &cpParameters.allowParameters);
     cpParameters.numParameter = (uint8 *)&(set.sig_parameter[WAVE_FORM]);
 }
